Fixed issues with the Terason scanner
git-svn-id: svn+ssh://cxserver.sintef.no/svn/Repository/CustusX3@1049 bf23512e-66b2-487e-abc1-aab7af62561f

Former-commit-id: 885ee8ff8c4b3093d2b57388db92d5d41e5e33cc

+++ modules/cxCommon/probeXmlConfigParser.h
@@ -23,9 +23,9 @@ public:
     std::string             mUsProbe;    ///<  probe
     std::string             mRtSource;   ///<  realtime source
     std::string             mConfigId;   ///<  config id
-    int                     mWidthDeg;   ///<  width
+    float                   mWidthDeg;   ///<  width
     int                     mDepth;      ///<  depth
-    int                     mOffset;     ///<  Offset
+    float                   mOffset;     ///<  Offset
     float                   mOriginCol;  ///<  Origin.Col
     float                   mOriginRow;  ///<  Origin.Row
     int                     mNCorners;   ///<  number of corners
